+++ test/jest/circleElement.test.js
@@ -417,10 +417,10 @@ describe("circleElement", () => {
     el.$options.hoverShowTimeout = 0;
     el.$refItems.children[1].dispatchEvent(new MouseEvent("mouseleave", { bubbles: true }));
     await h.wait(el.$options.hoverHideTimeout);
-    expect(el.querySelector("wup-popup").$isHidding).toBe(true);
+    expect(el.querySelector("wup-popup").$isHiding).toBe(true);
     el.$refItems.children[1].dispatchEvent(new MouseEvent("mouseenter", { bubbles: true }));
     await h.wait();
-    expect(el.querySelector("wup-popup").$isHidding).toBe(false);
+    expect(el.querySelector("wup-popup").$isHiding).toBe(false);
     expect(el.querySelector("wup-popup").$isShown).toBe(true);
 
     // new items